[CBRD-21142] Compile CUBRID code using C++ compiler (#719)
[CBRD-21142](http://jira.cubrid.org/browse/CBRD-21142)

I have modified CMakeLists.txt files so that almost all CUBRID code is now compiling using a C++ compiler, both on Linux and Windows.
For that I had to also change code because C++ is more strict with type casting, especially from int to enum and from void*.
There were also situations which I considered needing special attention - I tagged the lines with:
//vapa!!!
Here follows a list of problems found:
 * casts from int to enums
 * casts from void*
 * casts  void* <-> function pointer
 * incremmenting enums
 * initializing enum values with 0 (zero), while 0(zero) not being part of the enum.
 * casting function pointers with one signature to function pointers with another signature.

+++ src/query/crypt_opfunc.c
@@ -277,7 +277,7 @@ crypt_aes_default_encrypt (THREAD_ENTRY * thread_p, const char *src, int src_len
     }
   else
     {
-      padding_src_len = ceil ((double) src_len / AES128_BLOCK_LEN) * AES128_BLOCK_LEN;
+      padding_src_len = (int) ceil ((double) src_len / AES128_BLOCK_LEN) * AES128_BLOCK_LEN;
       pad = padding_src_len - src_len;
     }
